requirements.txt
streamlit>=1.28.0
openai>=1.13.3
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
pypdf>=3.17.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.5.0
plotly>=5.17.0

